+++ src/components/Drawer/components/UserInfo/UserInfo.tsx
@@ -124,7 +124,11 @@ const UserInfo: React.FC = () => {
       {DynamicChip}
       {/* TODO: Add tooltip explaining what an accountName is for */}
       <AccountNameChip size="small" label={accountName} onClick={() => displayAccountInfo()} />
-      <AccountBalanceChip size="small" label={NQTtoNXT(new BigNumber(balance), LongUnitPrecision) + " JUP"} onClick={() => handleCopy(balance)} />
+      <AccountBalanceChip
+        size="small"
+        label={NQTtoNXT(new BigNumber(balance), LongUnitPrecision) + " JUP"}
+        onClick={() => handleCopy(NQTtoNXT(new BigNumber(balance), LongUnitPrecision))}
+      />
     </>
   );
 };