+++ RELEASE-NOTES.md
@@ -14,3 +14,13 @@
 - Add ability to send JUP directly from the address book component
 - Displays the last block time on the Generators page in 24hr time format instead of 12hr
 - Remove My Accounts from gear menu
+
+## 0.0.4
+
+- Fix balance copy precision issue
+- Improves styling of Dex page (search component & asset details)
+- Adds loading indicator to search box on Dex page & Dex widget
+- Fixes Dex search box not displaying results after valid search
+- Fix asset precision/quantity issue by properly converting assets with different decimals params
+- Improve text on DEX page when no Asset is selected
+- Add icon and text to address book add button